Began implementing abbreviations for Auto Splitters
The code is becoming a bit unreadable with all the LINQ going on.
Also there seem to be a lot of games with a colon, like
A Hat in Time: Beta Build, where it abbreviates it to just
Beta Build which is really dumb. So, I'm not sure if we even want
this, as the results are not that great.