Added proper Backwards Compatability
If a Settings File from 1.5 or earlier is opened, LiveSplit now
automatically opens all of the split files in the history and tries
to figure out their Game and Category Names for the new Open Splits
section. This makes LiveSplit load up a little slower, but is only
happening once for each user that upgrades from 1.5 or earlier.