Improved Variables with Yes and No as Values
Variables that only have the Values "Yes" and "No" don't provide
any useful information in the extended category name, which is why
they are displayed differently now. For example, if the Variable is
about whether DLC is used, it'll show "DLC" and "No DLC" instead,
which is a lot more meaningful. This breaks if the Variable Name is
actually a question, but you can't do a lot about that.