Begin implementing Drift Compensation
By synchronizing with a Time Server, one can figure out what the drift of
the internal timer is. If this is done over a long interval, you can
approximate this really well without having a lot of issues with ping and
other internet related issues. By storing this into the settings.cfg,
which is not done as of this commit, you can improve the internal timer
more and more until it more or less runs as fast as an atomic clock.