Implement an Offline Safe Extended Category Name
Extended Category Names used to lock up if they request the Platform or
the Region. Instead they now check if those are already available at that
point in time. It'll use a simplified version of those, if the method is
not supposed to wait, which is a new parameter.

Fixes #388