Search speedrun.com in the background
This was previously being done on the main thread. Because it's slow to
update the tree, it locked the GUI while it updated everything. This
prevented the user from being able to interact with anything until the
task was complete.

Running it in the background means the GUI doesn't get locked, and also
means a new search can be run cancelling the old one before it is
complete.