Update Spyro Reignited Trilogy Autosplitter
With the approval of the leaderboard moderators, I have created an update to the Spyro Reignited Trilogy autosplitter, resolving many issues that the original autosplitter had. Changelog can be found in the ASL file, however changes include: 

- SRT3, Sorceress Last Hit fixed and enabled
- SRT3, Buzz, Spike, Scorch autosplit logic added
- SRT3, Sgt Byrd's Base autosplit fixed, typo in the level name
- SRT1, Autosplit added to Balloonists, with or without an EBL (required differing logic)

There will likely be further changes, however at this stage these QOL fixes are ready to be shipped.