Get SRC categories using the game ID, not the name
This commit changes the way the categories for a game on speedrun.com are retrieved.
Previously, it would try searching using the name which works in most cases but there are some edge cases that are mishandled by the speedrun.com API.
Now, the categories will be retrieved using the game ID which will be previously cached by the CompositeGameList inside a dictionary.
This change only adds a field and 2 methods to the CompositeGameList class which shouldn't break applications relying on the previous behavior.