Adds HCPComparisonGenerator (#2587)
* Adds HCPComparisonGenerator

HCP is a way of calculating results in Golf. It's based of the last 20
rounds played and averaging the top 8 results in that set.

This is an application of this method to give runners a goal different
from beating their PBs, which tends to be the case for a lot of runners.
This new HCP goal aims for the runner to be able to improve their
overall timings, or handicap as it's commonly referred to, without the
common mental pressure of beating their best time ever, which can
demotivate a lot of folks when they don't see their skills evolving.

* Rename variables for more descriptive ones

* Use Linq Average for calculation